requirements.txt
Cython==0.29.33
numpy==2.1.1
scikit-learn==1.5.1
web3==6.0.0
python-dotenv==1.0.0
joblib==1.2.0
Flask==2.3.2
aiohttp==3.10.5
pandas==1.5.3
setuptools>=65.5.1
setuptools-scm>=6.4.2